Internal: Address PR review comments [ED-24890]
- Remove explanatory code comments in app-bar.tsx per review feedback.
- Consolidate all editor-app-bar constants (MIN_APP_BAR_WIDTH,
  DEFAULT_MAX_TOOLBAR_ACTIONS, MaxToolbarActions) into constants.ts instead
  of splitting them across constants.ts and app-bar-size-context.tsx.
